solved precision problem: time variables TIME and DT from double to long double

+++ laser.cpp
@@ -2,6 +2,7 @@
 #include"electric_field.h"
 #include<iostream>
 #include<fstream>
+#include<math.h>
 
 using std::cout;
 using std::endl;
@@ -16,12 +17,13 @@ void laser::Init()
         K = 2.0*PI/LAMBDA;
 }
 
-electric_field laser::GetField(double d, double t)
+electric_field laser::GetField(double d, long double t)
 {
-	double phi, In;
+	long double phi, In;
 	electric_field ef;
 	//phase of the emitted field when it reaches x at time t
-	phi = K*d - OMEGA*(t - d/(2.0*C)) + oscillator(t - d/(2.0*C));
+	phi = K*d - OMEGA*(t - d/(2.0*C)) 
+		+ oscillator(t - d/(2.0*C));
 	//intensity of the field
 	In = INTENSITY;
 	ef.SetA(In);
